LayoutTests/fast/table/table-create-tbody.html

<!DOCTYPE html>
<script src="../../resources/dump-as-markup.js"></script>
<p>https://bugs.webkit.org/show_bug.cgi?id=84465 : Implement createTBody for table element.</p>
<p>This test verify that createTBody() appends a new tbody tag at the end of the table.</p>
<table>
  <thead>
  </thead>
  <tfoot>
  </tfoot>
</table>
<script>
var table = document.getElementsByTagName("table")[0];
table.createTBody();
Markup.dump(table);
</script>
