Safari-611.1.10/LayoutTests/fast/table/table-create-tbody-expected.txt

| "\n  "
| <thead>
|   "\n  "
| "\n  "
| <tfoot>
|   "\n  "
| "\n"
| <tbody>
